Bench: 7072296

+++ Halogen/src/Network.h
@@ -23,7 +23,7 @@ struct trainingPoint
 struct deltaPoint
 {
     size_t index;
-    float delta;
+    int delta;
 };
 
 struct Neuron
